+++ src/stac_asset/client.py
@@ -4,16 +4,22 @@ import asyncio
 import os.path
 import warnings
 from abc import ABC, abstractmethod
+from asyncio import Task
 from pathlib import Path
 from types import TracebackType
-from typing import AsyncIterator, Optional, TypeVar
+from typing import Any, AsyncIterator, Optional, TypeVar
 
 import aiofiles
 import pystac.utils
 from pystac import Item
 from yarl import URL
 
-from .errors import AssetDownloadException, AssetDownloadWarning
+from .errors import (
+    AssetDownloadException,
+    AssetDownloadWarning,
+    AssetOverwriteException,
+)
+from .strategy import FileNameStrategy
 from .types import PathLikeObject
 
 T = TypeVar("T", bound="Client")
@@ -78,6 +84,7 @@ class Client(ABC):
         make_directory: bool = False,
         item_file_name: Optional[str] = None,
         include_self_link: bool = True,
+        asset_file_name_strategy: FileNameStrategy = FileNameStrategy.FILE_NAME,
     ) -> Item:
         """Downloads all Assets in an item into the given directory.
 
@@ -95,6 +102,9 @@ class Client(ABC):
         directory_as_path = Path(directory)
         if make_directory:
             directory_as_path.mkdir(exist_ok=True)
+        elif not directory_as_path.exists():
+            raise FileNotFoundError(f"output directory does not exist: {directory}")
+
         if item_file_name is None:
             item_file_name = f"{item.id}.json"
         item_path = directory_as_path / item_file_name
@@ -105,17 +115,36 @@ class Client(ABC):
             except Exception as e:
                 raise AssetDownloadException(key, href, e)
 
-        tasks = list()
+        tasks: list[Task[Any]] = list()
+        keys_to_delete = list()
+        file_names: dict[str, str] = dict()
         for key, asset in item.assets.items():
-            # TODO allow different layout schemes
-            path = directory_as_path / os.path.basename(URL(asset.href).path)
+            if asset_file_name_strategy == FileNameStrategy.FILE_NAME:
+                file_name = os.path.basename(URL(asset.href).path)
+            elif asset_file_name_strategy == FileNameStrategy.KEY:
+                file_name = key + Path(asset.href).suffix
+            path = directory_as_path / file_name
+            if file_name in file_names:
+                raise AssetOverwriteException(list(file_names.values()))
+            else:
+                file_names[file_name] = str(path)
+
             absolute_href = asset.get_absolute_href()
             if absolute_href is None:
-                raise ValueError(f"asset '{key}' does not have an absolute href")
-            tasks.append(asyncio.create_task(download_asset(key, absolute_href, path)))
-            item.assets[key].href = pystac.utils.make_relative_href(
-                str(path), str(item_path)
-            )
+                warnings.warn(
+                    f"asset '{key}' does not have an absolute href",
+                    AssetDownloadWarning,
+                )
+                keys_to_delete.append(key)
+            else:
+                tasks.append(
+                    asyncio.create_task(download_asset(key, absolute_href, path))
+                )
+                item.assets[key].href = pystac.utils.make_relative_href(
+                    str(path), str(item_path)
+                )
+        for key in keys_to_delete:
+            del item.assets[key]
 
         results = await asyncio.gather(*tasks, return_exceptions=True)
         for result in results: